pyproject.toml
[project]
name = "openjwc-webapi"
version = "0.1.0"
description = "监听端口，给出响应"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "chromadb>=1.5.5",
    "fastapi>=0.135.1",
    "httpx[socks]>=0.28.1",
    "openai>=2.26.0",
    "passlib>=1.7.4",
    "pyjwt>=2.8.0",
    "rich>=14.3.3",
    "uvicorn>=0.41.0",
    "zhipuai>=2.1.5.20250825",
]
